chore: rename agents to be consistent
