+++ mesh/tests/bitquery_sol_token_info_agent.py
@@ -8,7 +8,7 @@ import asyncio
 # Add the project root to the Python path
 sys.path.append(str(Path(__file__).parent.parent.parent))
 
-from mesh.bitquery_sol_token_info_agent import BitquerySolanaTokenInfoAgent
+from mesh.bitquery_solana_token_info_agent import BitquerySolanaTokenInfoAgent
 
 load_dotenv()
